Move DatStorer to storers
